==> src/psyberwatch/notebooks/0. Environment Setup.ipynb <==
# Project Environment Setup
To improve the environment experience, the below cells will execute a series of bash scripts to install all required packages and dependencies. The base python installed on the machine (Python that is on PATH) can be used for this notebook.

Create virtual environments.

!python3 -m venv ./environments/logit/.logit-venv
!python3 -m venv ./environments/logit/.svm-venv

Activate environment, install packages, and deactivate

import platform

system = platform.system()

if system == "Windows":
    !source ./environments/logit/.logit-venv/Scripts/activate && pip3 install -r ./environments/logit/requirements.txt && deactivate
    !source ./environments/svm/.svm-venv/bin/Scripts && pip3 install -r ./environments/svm/requirements.txt && deactivate
elif system in ["Darwin", "Linux"]:
    !source ./environments/logit/.logit-venv/bin/activate && pip install -r ./environments/logit/requirements.txt && deactivate
    !source ./environments/svm/.svm-venv/bin/activate && pip install -r ./environments/svm/requirements.txt && deactivate
