==> src/psyberwatch/notebooks/1. Data Collection.ipynb <==
# Data Collection
---
Ham and spam data is downloaded from [marcelwiechmann/enron-spam-data](https://www.kaggle.com/datasets/marcelwiechmann/enron-spam-data) and the phishing data is downloaded from [subhajournal/phishingemails](https://www.kaggle.com/datasets/subhajournal/phishingemails).

## 1. Imports

import pandas as pd
import kagglehub
import os
from pathlib import Path

## 2. Download the Data from Kaggle
Download the datasets using `kagglehub`, and are stored in `/home/<user>//.cache/kagglehub/datasets/...` (Unix/Linux path). Path may be different if installed on Windows system.

Create `data` folder.

Path("../data").mkdir(parents=True, exist_ok=True)

Download ham and spam data.

Alternative datasets:
- marcelwiechmann/enron-spam-data
- jackksoncsie/spam-email-dataset
- subhajournal/phishingemails

from_source = "meruvulikith/190k-spam-ham-email-dataset-for-classification"
full_local_path = f"{os.path.expanduser('~')}/.cache/kagglehub/datasets/meruvulikith/190k-spam-ham-email-dataset-for-classification/versions/1/"

if not Path(full_local_path).exists():
    # If the file doesn't exist, download it
    full_local_path = kagglehub.dataset_download(from_source)

hamSpam_df = pd.read_csv(f"{full_local_path}/spam_Emails_data.csv")
hamSpam_df.to_csv("../data/1_hamSpam.csv", index=False)
hamSpam_df.head(5)

hamSpam_df["label"].value_counts()

Number of true duplicates

hamSpam_df[hamSpam_df.duplicated() == True]["label"].value_counts()

Download phishing data.

full_local_path = f"{os.path.expanduser('~')}/.cache/kagglehub/datasets/naserabdullahalam/phishing-email-dataset/versions/1/"

if not Path(full_local_path).exists():
    path = kagglehub.dataset_download("subhajournal/phishingemails")

# Define the list of files to read
file_names = ["Ling.csv", "Nazario.csv", "Nigerian_Fraud.csv", "CEAS_08.csv"]

# List to hold individual dataframes
dataframes = []

# Loop through each file, read it, and append it to the list
for file_name in file_names:
    file_path = os.path.join(full_local_path, file_name)
    try:
        df = pd.read_csv(file_path)
        dataframes.append(df)
        print(f"Successfully loaded: {file_name}")
    except FileNotFoundError:
        print(f"Error: File not found at {file_path}")
    except Exception as e:
        print(f"An error occurred while reading {file_name}: {e}")

phish_df = pd.concat(dataframes, ignore_index=True)
phish_df = phish_df[phish_df["label"] == 1]
phish_df.to_csv("../data/1_phish.csv", index=False)
phish_df.head(5)

print("Number of true duplicates: ", phish_df[phish_df.duplicated() == True]["label"].value_counts())

Show the columns of each dataset.

list(hamSpam_df.columns.values) , list(phish_df.columns.values)

Show the current count of each dataset.

hamSpam_df.shape, phish_df.shape

Data from the two sources are not compatible at initial load. We will have to clean the data and combine them in a common dataframe structure. Continue to step 2.
