==> src/psyberwatch/notebooks/2. Data Cleaning.ipynb <==
# Data Cleaning
---
Take the raw data from `/data/hamSpam.csv` and `/data/phish.csv` and conform the ham/spam dataset to the structure of the phishing dataset and integrate together. Then clean the dataset by removing NA values, standardize the textual features by lowercasing and removing stopwords, and remove any embedded HTML elements from the text.

## 1. Imports & Setup

import pandas as pd

import nltk
nltk.download('stopwords')
from nltk.tokenize import word_tokenize

from src.utils import clean_text

Download nltk datasets.

# NLTK downloads
nltk.download("punkt")
nltk.download("wordnet")
nltk.download('punkt_tab')

## 2. Load Data from CSV

hamSpam_df = pd.read_csv("../data/1_hamSpam.csv")
phish_df = pd.read_csv("../data/1_phish.csv")

## 3. Process Ham/Spam Data

# Rename known columns (no error if source names are missing)
hamSpam_df.rename(
    columns={"label": "Email Type", "text": "Email Text"},
    inplace=True
)

hamSpam_df["Email Type"] = hamSpam_df["Email Type"].str.lower()

# Export as processed dataset
hamSpam_df.to_feather(
    "../data/2_hamSpam_processed.feather"
)

## 4. Process Phishing Data

Observe email types.

phish_df["label"].apply(repr).unique()

Standardize email types.

other_columns = ["subject", "sender", "receiver", "date", "urls"]
if set(other_columns).intersection(phish_df.columns) == set(other_columns):
    phish_df.drop(columns=["subject", "sender", "receiver", "date", "urls"], inplace=True)

phish_df.rename(
    columns={"label": "Email Type", "body": "Email Text"},
    inplace=True
)

phish_df['Email Type'] = phish_df['Email Type'].replace(
    {1: 'phish'}
)

phish_df['Email Type'].apply(repr).unique()
phish_df

## 5. Combine Datasets

email_df = pd.concat([hamSpam_df, phish_df], ignore_index=True)
email_df['Email Type'].apply(repr).unique()

Display examples of each category.

email_df[email_df["Email Type"] == "spam"].iloc[[5000]]

email_df[email_df["Email Type"] == "ham"].iloc[[5000]]

email_df[email_df["Email Type"] == "phish"].iloc[[5000]]

Show total count of emails.

email_df.count()

Show counts for each category.

email_df["Email Type"].value_counts()

## 6. Drop NA and "empty" Rows

print(email_df.isna().sum())

email_df.dropna(inplace=True)

empty_rows = email_df[email_df["Email Text"] == "empty"]
email_df.drop(empty_rows.index, inplace=True)

print(email_df.isna().sum())

## 7. Drop Any Duplicate Rows
`spam` contains many duplicated rows, which will cause the model to overfit to spam email data.

email_df[email_df["Email Type"] == "spam"][email_df.duplicated() == True].count()

email_df[email_df["Email Type"] == "ham"][email_df.duplicated() == True].count()

email_df[email_df["Email Type"] == "phish"][email_df.duplicated() == True].count()

Drop any duplicates and confirm that count of duplicates is 0.

email_df.drop_duplicates(keep="first", inplace=True)
email_df[email_df.duplicated() == True].count()

Get total count of dataset after removing duplicates

email_df["Email Type"].value_counts()

## 7. Bulk Clean Dataset & Export to CSV


Clean data by removing stop words, punctuation, special characters, and tokenize/lemmatize.

cleaned_email_df = email_df.copy()
cleaned_email_df["Cleaned Email Text"] = email_df["Email Text"].apply(clean_text)

Export to feather.

cleaned_email_df.to_feather(
    "../data/2_clean_email_dataset.feather"
)
