==> src/psyberwatch/notebooks/5. Demo Classify Emails.ipynb <==
# Demonstrate the Classification of new Emails using Existing Model
---
TODO:
- Use the same pre-processing/cleaning functions used before training the model and apply it to sample email.
- Use joblib to save model for future use.

## Imports

import joblib
import pandas as pd
from utils import clean_text

## Load Model

model = joblib.load("./models/svm_model.joblib")

## Clean & Encode Test Email

email_dataframe = pd.read_feather("./data/2_clean_email_dataset.feather")

mapping = dict(zip(
    email_dataframe["Email Type"].astype("category").cat.codes.unique(), email_dataframe["Email Type"].unique()
))

# Retrieve device variable from 4. Analyze using SVM.ipynb
%store -r device

sbert = joblib.load("./models/sbert.joblib")

def encode_email(email_text):
    cleaned = clean_text(email_text)
    embedding = sbert.encode(
        cleaned,
        convert_to_numpy=True
    ).reshape(1, -1)
    return embedding

def classify_email(email_text):
    embedding = encode_email(email_text)
    prediction = model.predict(embedding)[0]
    return mapping[prediction]

## Test

test_email = """
URGENT: Your account has been compromised. 
Click the link below to verify your identity immediately.
"""

result = classify_email(test_email)
print("Predicted class:", result)

test_email = """
Hey steve, I'm sorry to hear that you're having trouble with your account. 
I'll do my best to help you out.
"""

result = classify_email(test_email)
print("Predicted class:", result)

test_email = """
Enter the code below to verify your identity:
"""

result = classify_email(test_email)
print("Predicted class:", result)

test_email = """
You've been selected for a special promotion! 
Click here to claim your prize:
"""

result = classify_email(test_email)
print("Predicted class:", result)

test_email = """
Are you in need of a new bike? Check out our latest collection:
"""

result = classify_email(test_email)
print("Predicted class:", result)
